Revert

+++ src/transformers/models/zamba/modeling_zamba.py
@@ -170,7 +170,7 @@ class ZambaAttention(nn.Module):
         )
 
         attn_output = attn_output.reshape(*input_shape, -1).contiguous()
-        attn_output = self.o_proj(attn_output.to(hidden_states.dtype))
+        attn_output = self.o_proj(attn_output)
         return attn_output, attn_weights
